requirements.txt
ansible==1.6.10
boto==2.32.0
paramiko==1.14.0
python-decouple==2.2
